-fallback: builtin_predicates.py FormularVisitor.py formular_manager.py -minor changes for 3d adaptive verification: map_verification.py map_verification_repairing.py -modified : add LLM's content to the exsiting code: formula_ids.py groups_handler.py formula_collection.py lanelet_predicates.py
