+++ crdesigner/verification_repairing/verification/formula_ids.py
@@ -5,16 +5,6 @@ from typing import List, Union
 # type of the CommonRoad element.
 
 @enum.unique
-class LaneletFormulaID3D(enum.Enum):
-    POLYLINES_INTERSECTION_3D   = "polylines_intersection_3d"
-    LEFT_SELF_INTERSECTION_3D   = "left_self_intersection_3d"
-    RIGHT_SELF_INTERSECTION_3D  = "right_self_intersection_3d"
-    LANELET_VERTICAL_CLEARANCE  = "lanelet_vertical_clearance"
-    VERTICAL_CLEARANCE_STACKED  = "vertical_clearance_stacked"
-    GRADE_WITHIN_LIMIT          = "grade_within_limit"
-    PREDECESSOR_VERTICAL_STEP   = "predecessor_vertical_step"
-    TUNNEL_DEPTH_VALID          = "tunnel_depth_valid"
-@enum.unique
 class GeneralFormulaID(enum.Enum):
     """The IDs of formulas that describe the properties of all types of elements."""
 
@@ -72,6 +62,14 @@ class LaneletFormulaID(enum.Enum):
     CONFLICTING_LANELET_DIRECTIONS = "conflicting_lanelet_directions"
     LEFT_RIGHT_BOUNDARY_ASSIGNMENT = "left_right_boundary_assignment"
 
+    POLYLINES_INTERSECTION_3D   = "polylines_intersection_3d"
+    LEFT_SELF_INTERSECTION_3D   = "left_self_intersection_3d"
+    RIGHT_SELF_INTERSECTION_3D  = "right_self_intersection_3d"
+    LANELET_VERTICAL_CLEARANCE  = "lanelet_vertical_clearance"
+    VERTICAL_CLEARANCE_STACKED  = "vertical_clearance_stacked"
+    GRADE_WITHIN_LIMIT          = "grade_within_limit"
+    PREDECESSOR_VERTICAL_STEP   = "predecessor_vertical_step"
+
 
 @enum.unique
 class TrafficSignFormulaID(enum.Enum):
@@ -112,7 +110,6 @@ class IntersectionFormulaID(enum.Enum):
 
 FormulaID = Union[
     LaneletFormulaID,
-    LaneletFormulaID3D,
     TrafficSignFormulaID,
     TrafficLightFormulaID,
     IntersectionFormulaID,
@@ -120,7 +117,6 @@ FormulaID = Union[
 ]
 FormulaTypes = [
     LaneletFormulaID,
-    LaneletFormulaID3D,
     TrafficSignFormulaID,
     TrafficLightFormulaID,
     IntersectionFormulaID,
