+++ crdesigner/verification_repairing/verification/hol/formula_manager.py
@@ -1,7 +1,5 @@
-from __future__ import annotations
-
 import warnings
-from typing import Any, Dict, List, Set, Optional
+from typing import Any, Dict, List, Set
 
 from crdesigner.verification_repairing.verification.hol.formula import Formula
 from crdesigner.verification_repairing.verification.hol.formula_collection import (
@@ -10,95 +8,76 @@ from crdesigner.verification_repairing.verification.hol.formula_collection impor
     LaneletFormulas,
     TrafficLightFormulas,
     TrafficSignFormulas,
-    LaneletFormulas3D,
 )
 from crdesigner.verification_repairing.verification.hol.parser.parser import Parser
 
 
 class FormulaManager:
     """
-    Central registry that converts all textual HOL‑specifications into
-    executable expression trees and stores accompanying domains.
+    Class representing the management of formulas.
     """
 
-    # ────────────────────────────── ctor ──────────────────────────────
-    def __init__(self, lanelet_network: Optional["LaneletNetwork"] = None) -> None:
+    def __init__(self):
         """
-        :param lanelet_network: CommonRoad LaneletNetwork.  
-                               Used only to detect whether 3‑D rules are required.
+        Constructor.
         """
-        # True: # At least one lanelet has a third dimension -> enable 3D rules
-        self._enable_3d: bool = (
-            lanelet_network is not None
-            and any(ll.center_vertices.shape[1] == 3 for ll in lanelet_network.lanelets)
-        )
-
-        self._formulas: List[Formula] = []
-        self._domains: Dict[str, Set[Any]] = {}
+        self._formulas = []
+        self._domains = {}
 
         self._collect_formulas()
 
-    # ─────────────────────────── properties ───────────────────────────
     @property
     def formulas(self) -> List[Formula]:
         return self._formulas
 
     @formulas.setter
-    def formulas(self, value: List[Formula]) -> None:
-        self._formulas = value
+    def formulas(self, formulas: List[Formula]):
+        self._formulas = formulas
 
     @property
     def domains(self) -> Dict[str, Set[Any]]:
         return self._domains
 
     @domains.setter
-    def domains(self, value: Dict[str, Set[Any]]) -> None:
-        self._domains = value
+    def domains(self, domains: Dict[str, Set[Any]]):
+        self._domains = domains
 
-    # ────────────────────── public helper api ────────────────────────
-    def add_formula(self, formula: Formula) -> None:
-        """Add a new formula unless its ID already exists."""
-        if any(f.formula_id == formula.formula_id for f in self._formulas):
-            warnings.warn(f"Formula with ID {formula.formula_id} is already stored!")
-            return
+    def add_formula(self, formula: Formula):
+        """
+        Adds a formula. A formula is not stored if a formula with the same ID is already contained.
+
+        :param formula: Formula.
+        """
+        for f in self._formulas:
+            if f.formula_id == formula.formula_id:
+                warnings.warn("Formula with ID {} is already stored!".format(formula.formula_id))
+                return
         self._formulas.append(formula)
 
-    def add_domain(self, domain_id: str, values: Set[Any]) -> None:
-        """Register a new domain unless it has been defined before."""
-        if domain_id in self._domains:
-            warnings.warn(f"Domain with ID {domain_id} is already stored!")
-            return
-        self._domains[domain_id] = set(values)
+    def add_domain(self, domain_id: str, values: Set[Any]):
+        """
+        Adds a domain. A domain is not stored if a domain with the same ID is already contained.
 
-    # ─────────────────────── internal routines ───────────────────────
-    def _collect_formulas(self) -> None:
+        :param domain_id: Domain ID.
+        :param values: Values.
         """
-        Convert every formula string from every collection into
-        executable expression trees and gather their domains.
-        """
-        # 1. decide which collections to load
-        collections = [
+        if domain_id in self._domains.keys():
+            warnings.warn("Domain with ID {} is already stored!".format(domain_id))
+            return
+        self._domains[domain_id] = values
+
+    def _collect_formulas(self):
+        for collection in [
             TrafficLightFormulas,
             TrafficSignFormulas,
             IntersectionFormulas,
             LaneletFormulas,
             GeneralFormulas,
-        ]
-        if self._enable_3d:
-            # Place 3D formulas first, to detect them early
-            collections.insert(0, LaneletFormulas3D)
+        ]:
+            for formula_id, formula in collection.formulas.items():
+                for subformula_id, subformula in collection.subformulas.items():
+                    formula = formula.replace(subformula_id, subformula)
+                self._formulas.append(Parser.parse(formula, formula_id))
 
-        # 2. parse all formulas & merge sub‑formulas
-        for coll in collections:
-            for formula_id, formula_text in coll.formulas.items():
-                # substitute sub‑formula placeholders
-                for sub_id, sub_text in coll.subformulas.items():
-                    formula_text = formula_text.replace(sub_id, sub_text)
-
-                parsed = Parser.parse(formula_text, formula_id)
-                self._formulas.append(parsed)
-
-            # merge domains (later collections can overwrite earlier ones,
-            # which is useful for specialised versions)
-            for dom_id, dom_vals in coll.domains.items():
-                self._domains.setdefault(dom_id, set()).update(dom_vals)
+            for domain_id, values in collection.domains.items():
+                self._domains[domain_id] = set(values)
